add article screen

+++ routes/file.js
@@ -60,13 +60,54 @@ router.post('/files/:filename', upload.single('file'), function (req, res) {
     var query = { 'user_id': user_id, 'GalleryName': GalleryName, 'filename': filename, 'gallery_id': gallery_id };
     collection.find(query).toArray(function (_err, docs) {
       if (docs.length === 0) {
-        query = { 'image_id': uniqid.time(), 'user_id': user_id, 'gallery_id': gallery_id, 'GalleryName': GalleryName, 'filename': filename, 'imageStatus':imageStatus };
+        query = { 'image_id': uniqid.time(), 'user_id': user_id, 'gallery_id': gallery_id, 'GalleryName': GalleryName, 'filename': filename, 'imageStatus': imageStatus };
         collection.insertOne(query, function (_err, inserted) {
           // res.json(_displayResults(_resultCode.IMAGE_UPLOADED_SUCCESS, 'Successfully uploaded', true));
         });
       } else {
         // res.status(201).json(_displayResults(_resultCode.IMAGE_ALREADY_EXIST, 'This image already exist'));
       }
+    });
+    res.json(_displayResults(_resultCode.FILE_UPLOAD_SUCCESS, 'Successfully uploaded', true));
+  } else {
+    res.status(405).json(_displayResults(_resultCode.NO_SELECTED_UPLOAD_FILE, "Please select file(key='file') to upload"));
+  }
+});
+
+var storage = multer.diskStorage({
+  destination: (req, file, cb) => {
+    var article_id = req.body.article_id;
+    var dir = 'uploads/articles/' + article_id;
+    console.log(dir);
+    if (!fs.existsSync(dir)) {
+      fs.mkdirSync(dir);
+    }
+    cb(null, dir);
+  },
+  filename: (req, file, cb) => {
+    var filename = req.params.filename;
+    if (filename === undefined) {
+      filename = menufile;
+    }
+    cb(null, filename);
+  }
+});
+
+var maxSize = config.upload_file_max_size;
+var upload = multer({ storage: storage, limits: { fileSize: maxSize } });
+
+/** ************************ file uplaod and download route *************************/
+/* file upload : POST */
+router.post('/files/otherarticleimg/:filename', upload.single('file'), function (req, res) {
+  if (req.file) {
+    var article_id = req.body.article_id;
+    var filename = req.params.filename;
+
+    var collection = db.get().collection(config.articleimagesCollection);
+    collection.find().toArray(function (_err, docs) {
+      query = { 'image_id': uniqid.time(), 'article_id': article_id, 'filename': filename };
+      collection.insertOne(query, function (_err, inserted) {
+      });
     });
     res.json(_displayResults(_resultCode.FILE_UPLOAD_SUCCESS, 'Successfully uploaded', true));
   } else {
@@ -243,15 +284,19 @@ router.get('/files', function (req, res) {
 // });
 
 /* Create article : POST */
+var articleId
 
 var storage = multer.diskStorage({
   destination: (req, file, cb) => {
     if (!fs.existsSync('uploads')) {
       fs.mkdirSync('uploads');
     }
-
-
     var dir = 'uploads/articles';
+    if (!fs.existsSync(dir)) {
+      fs.mkdirSync(dir);
+    }
+    articleId = uniqid.time();
+    var dir = 'uploads/articles/' + articleId;
     if (!fs.existsSync(dir)) {
       fs.mkdirSync(dir);
     }
@@ -262,6 +307,7 @@ var storage = multer.diskStorage({
     if (filename === undefined) {
       filename = menufile;
     }
+    filename = articleId + filename;
     cb(null, filename);
   }
 });
@@ -298,6 +344,7 @@ router.post('/getdata/addarticle/:filename', upload.single('file'), function (re
       res.status(405).json(_displayResults(_resultCode.MAINIMAGE_UNDEFINED, 'mainImage is undefined'));
       return;
     }
+    mainImage = articleId + mainImage;
     var sticky = req.body.sticky;
     if (sticky === undefined) {
       res.status(405).json(_displayResults(_resultCode.STICKY_UNDEFINED, 'sticky is undefined'));
@@ -308,7 +355,7 @@ router.post('/getdata/addarticle/:filename', upload.single('file'), function (re
     var query = { 'title': title };
     collection.find(query).toArray(function (_err, docs) {
       if (docs.length === 0) {
-        query = { 'article_id': uniqid.time(), 'title': title, 'content': content, 'status_b': status_b, 'dateTime': dateTime, 'mainImage': mainImage, 'sticky': sticky };
+        query = { 'article_id': articleId, 'title': title, 'content': content, 'status_b': status_b, 'dateTime': dateTime, 'mainImage': mainImage, 'othersImg': '', 'sticky': sticky };
         collection.insertOne(query, function (_err, inserted) {
           res.json(_displayResults(_resultCode.ARTICLE_CREATED_SUCCESS, 'Successfully created', true));
         });
@@ -322,6 +369,110 @@ router.post('/getdata/addarticle/:filename', upload.single('file'), function (re
 });
 
 
+/* Edit article : POST */
+var article_id_edit;
+var storage = multer.diskStorage({
+
+  destination: (req, file, cb) => {
+    if (!fs.existsSync('uploads')) {
+      fs.mkdirSync('uploads');
+    }
+    var dir = 'uploads/articles';
+    if (!fs.existsSync(dir)) {
+      fs.mkdirSync(dir);
+    }
+    article_id_edit = req.body.article_id;
+    var dir = 'uploads/articles/' + article_id_edit;
+    if (!fs.existsSync(dir)) {
+      fs.mkdirSync(dir);
+    }
+    cb(null, dir);
+  },
+  filename: (req, file, cb) => {
+    var filename = req.params.filename;
+    if (filename === undefined) {
+      filename = menufile;
+    }
+    filename = article_id_edit + filename;
+    cb(null, filename);
+  }
+});
+
+var maxSize = config.upload_file_max_size;
+var upload = multer({ storage: storage, limits: { fileSize: maxSize } });
+
+/** ************************ file uplaod and download route *************************/
+/* file upload : POST */
+router.post('/getdata/editarticle/:filename', upload.single('file'), function (req, res) {
+  var article_id = req.body.article_id;
+  if (article_id === undefined) {
+    res.status(405).json(_displayResults(_resultCode.TITLE_UNDEFINED, 'article_id is undefined'));
+    return;
+  }
+  var title = req.body.title;
+  if (title === undefined) {
+    res.status(405).json(_displayResults(_resultCode.TITLE_UNDEFINED, 'Title(title) is undefined'));
+    return;
+  }
+  var content = req.body.content;
+  if (content === undefined) {
+    res.status(405).json(_displayResults(_resultCode.CONTENT_UNDEFINED, 'Content(content) is undefined'));
+    return;
+  }
+  var status_b = req.body.status_b;
+  if (status_b === undefined) {
+    res.status(405).json(_displayResults(_resultCode.STATUS_UNDEFINED, 'Status_b(status_b) is undefined'));
+    return;
+  }
+  var dateTime = req.body.dateTime;
+  if (dateTime === undefined) {
+    res.status(405).json(_displayResults(_resultCode.DATETIME_UNDEFINED, 'dateTime is undefined'));
+    return;
+  }
+  var mainImage = req.body.mainImage;
+  if (mainImage === undefined) {
+    res.status(405).json(_displayResults(_resultCode.MAINIMAGE_UNDEFINED, 'mainImage is undefined'));
+    return;
+  }
+
+  var sticky = req.body.sticky;
+  if (sticky === undefined) {
+    res.status(405).json(_displayResults(_resultCode.STICKY_UNDEFINED, 'sticky is undefined'));
+    return;
+  }
+  var collection = db.get().collection(config.articleCollection);
+  if (req.file !== undefined) {
+    //////////////
+    mainImage = article_id + mainImage;
+    ////////////
+    var mainImageOld;
+    var query = { 'article_id': article_id };
+    collection.find(query).toArray(function (_err, docs) {
+      mainImageOld = docs[0].mainImage
+      var curPath = 'uploads/articles/' + article_id + '/' + mainImageOld;
+      if (!fs.lstatSync(curPath).isDirectory()) { // delete file
+        fs.unlinkSync(curPath);
+      }
+    });
+  }
+  setTimeout(() => {
+    collection.find().toArray(function (_err, docs) {
+      collection.findOneAndUpdate(
+        { 'article_id': article_id },
+        { '$set': { 'title': title, 'content': content, 'status_b': status_b, 'dateTime': dateTime, 'mainImage': mainImage, 'othersImg': '', 'sticky': sticky } },
+        function (err, doc) {
+          if (err) { // err: any errors that occurred
+            res.send(err);
+          } else {
+            res.send(docs);
+          }
+        }
+      );
+    });
+  }, 1500);
+});
+
+
 /* Get articles: Get */
 router.get('/getdata/articles', (req, res) => {
   var collection = db.get().collection(config.articleCollection);
@@ -330,24 +481,74 @@ router.get('/getdata/articles', (req, res) => {
   });
 });
 
+/* Get articleImages: Get */
+router.get('/getdata/articleImages', (req, res) => {
+  var collection = db.get().collection(config.articleimagesCollection);
+  collection.find().toArray((err, results) => {
+    res.send(results);
+  });
+});
+
 router.post('/getdata/deleteArticle', function (req, res) {
   var article_id = req.body.article_id;
   var filename = req.body.filename;
-  console.log(filename);
-  var curPath = 'uploads/articles' + '/' + filename;
-  if (!fs.lstatSync(curPath).isDirectory()) { // delete file
-    fs.unlinkSync(curPath);
+  var path = 'uploads/articles/' + '/' + article_id;
+  if (fs.existsSync(path)) {
+    fs.readdirSync(path).forEach((file, index) => {
+      const curPath = Path.join(path, file);
+      if (!fs.lstatSync(curPath).isDirectory()) { // delete file
+        fs.unlinkSync(curPath);
+      }
+    });
+    fs.rmdirSync(path);
   }
+
+  // if (!fs.lstatSync(curPath).isDirectory()) { // delete file
+  //   fs.unlinkSync(curPath);
+  // }
   var collection = db.get().collection(config.articleCollection);
   var query = { 'article_id': article_id };
   collection.find(query).toArray(function (_err, docs) {
     if (docs.length !== 0) {
       collection.deleteOne(query, function (_err, inserted) {
-        res.json(_displayResults(_resultCode.ARTICLE_DELETE_SUCCESS, 'Successfully deleted', true));
+
+        collection = db.get().collection(config.articleimagesCollection);
+        collection.find(query).toArray(function (_err, docs) {
+          if (docs.length !== 0) {
+            collection.deleteMany(query, function (_err, inserted) {
+              res.json(_displayResults(_resultCode.GALLERY_DELETE_SUCCESS, 'Successfully deleted', true));
+            });
+          } else {
+            res.status(201).json(_displayResults(_resultCode.GALLERY_ALREADY_DELETED, 'This gallery already delete'));
+          }
+        });
       });
     }
   })
 });
+
+/* Delete Image */
+
+router.post('/getdata/deleteArticleImage', function (req, res) {
+  var article_id = req.body.article_id;
+  var filename = req.body.filename;
+  var curPath = 'uploads/articles/' + '/' + article_id + '/' + filename;
+  if (!fs.lstatSync(curPath).isDirectory()) { // delete file
+    fs.unlinkSync(curPath);
+  }
+  var collection = db.get().collection(config.articleimagesCollection);
+  var query = {'article_id': article_id, 'filename': filename };
+  collection.find(query).toArray(function (_err, docs) {
+    if (docs.length !== 0) {
+      collection.deleteOne(query, function (_err, inserted) {
+        res.json(_displayResults(_resultCode.GALLERY_DELETE_SUCCESS, 'Successfully deleted', true));
+      });
+    } else {
+      res.status(201).json(_displayResults(_resultCode.GALLERY_ALREADY_DELETED, 'This gallery already delete'));
+    }
+  });
+});
+
 /** ************************ menu route *************************/
 /* menu.json file upload : POST */
 router.post('/menu', function (req, res) {
